internals: settle the remaining C sources into their final form
Normalise the rest of the C -- the device sink, the font, path, token,
matrix, dict and control operators, the interpreter core and the small
device back ends -- to the source form the tree ships. This is largely
function ordering and helper-shape alignment (the text-state accessor, for
one, now takes its graphics-state argument last). No behaviour change.

Source code generated by Claude (Anthropic).
Commit message generated by Claude (Anthropic).

+++ src/lib/xpost_save.c
@@ -169,8 +169,8 @@ unsigned xpost_save_ent_is_saved(Xpost_Memory_File *mem,
        sav.lev + 1 (see save_save_ent) so the "already backed up" marker
        cannot be confused with an object's birth stamp, which equals its
        birth count and is therefore <= sav.lev for anything protectable. */
-    return llev <= (unsigned int)sav.save_.lev ?
-        (tlev == (unsigned int)sav.save_.lev + 1) : 1;
+    return llev <= sav.save_.lev ?
+        (tlev == sav.save_.lev + 1u) : 1;
 }
 
 /* make a clone of ent, return new ent */
